requirements.txt
fastapi==0.111.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.7,<0.1.0
python-dotenv==1.0.0
boto3==1.34.0
pydantic-settings==2.0.3
openai>=1.25.0
pymupdf==1.26.3
pillow==11.3.0
pandas==2.3.1
xlsxwriter==3.2.5
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
# OCR com Gemini
google-generativeai==0.8.5

